requirements.txt
Flask
mysql-connector-python
flask-swagger-ui
flask-restful
flask-restx